+++ packages/db/migrations/0042_worthless_thor.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "mcp_oauth_clients" ADD COLUMN "code_verifier" text;